project-3-commodities-etl/requirements.txt
pandas
SQLAlchemy
python-dotenv
requests
entsoe-py
streamlit